Interface: Add subfunction is_file_open() to the multiplex
In order that the FreeDOS kernel can determine if it's acceptable
to delete or rename a file it must first know whether a file is
open. This patch creates a new subfunction int2f/0x10a6 that the
kernel may use for this purpose. A similar patch, with identical
subfunction numbering, has been applied to FDPP see
https://github.com/dosemu2/fdpp/commit/691721f1.