Build: Add on the fly line conversion for TC 2.01
Turbo C 2.01 doesn't properly process files with Unix line endings,
and in fact it ignores the content completely. This is a problem in
situations such as when checked out via git for compiling on Linux
using an emulator. We could override the automatic platform text
conversion using a .gitattributes file to force CRLF, but this might
make things awkward when cross compiling with Gcc-ia16. For now we
can just convert back and forth when using Turbo C 2.01.