support AMIS general uninstallers
AMIS function 02h will now disable the resident interrupt 2Fh
handler, reset all SFTs' SHARE field, and patch the old kernel's
flag if needed. This means that general uninstallers like the
REMOVE included in amislib by Ralf Brown will now do the right
thing, if the TSR can be removed.

However, if the removal fails then we are now left with a
disabled TSR. Therefore, in our own uninstaller, we check for
the interrupt hooks first so that a failed removal is more of
a side-effect free action.